Move more helpers into the new file
