+++ compiler/optimizations/propagateDomainConstness.cpp
@@ -24,8 +24,61 @@
 #include "CallExpr.h"
 #include "DecoratedClassType.h"
 #include "expr.h"
+#include "IfExpr.h"
 #include "stmt.h"
 #include "view.h"
+
+#include <set>
+
+static void establishDefinedConstWithIfExprs(Expr* e) {
+  if (CallExpr *initCall = toCallExpr(e)) {
+    if (initCall->isNamed("chpl__buildDomainExpr")) {
+      initCall->argList.last()->replace(new SymExpr(gTrue));
+      return;
+    }
+    else if (initCall->isNamed("chpl__distributed")) {
+      initCall->get(3)->replace(new SymExpr(gTrue));
+      return;
+    }
+  }
+  else if (IfExpr *initIf = toIfExpr(e)) {
+    establishDefinedConstWithIfExprs(initIf->getThenStmt()->body.head);
+    establishDefinedConstWithIfExprs(initIf->getElseStmt()->body.head);
+  }
+}
+
+void establishDefinedConstIfApplicable(DefExpr* defExpr,
+                                       std::set<Flag>* flags) {
+
+  if (defExpr->init != NULL) {
+    if (flags->size() == 1 && flags->count(FLAG_CONST)) {
+      establishDefinedConstWithIfExprs(defExpr->init);
+    }
+  }
+
+  if (defExpr->exprType != NULL) {
+    if (CallExpr *initType = toCallExpr(defExpr->exprType)) {
+      if (initType->isNamed("chpl__distributed")) {
+        initType->get(3)->replace(new SymExpr(gTrue));
+        return;
+      }
+      else if (initType->isNamed("chpl__buildArrayRuntimeType")) {
+        if (CallExpr *typeCall = toCallExpr(initType->get(1))) {
+          if (typeCall->isNamed("chpl__ensureDomainExpr")) {
+            if (CallExpr *buildDomExpr = toCallExpr(typeCall->get(1))) {
+              if (buildDomExpr->isNamed("chpl__buildDomainExpr")) {
+                buildDomExpr->argList.last()->replace(new SymExpr(gTrue));
+                return;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+                
+
 
 static VarSymbol *addFieldAccess(Symbol *receiver, const char *fieldName,
                                  Expr *insBefore, Expr *&insAfter,
